Migrationsv2: limit batch sizes to migrations.batchSizeBytes (= 100mb by default) (#109540)
* Fix logging for existing integration test

* First stab at limiting batches to batchSizeBytes

* Fix tests

* Fix batch size calculation, NDJSON needs to be terminated by an empty line

* Integration tests

* Fix type failures

* rename migration integration tests and log files to be consistent & more descriptive

* Review feedback

* Remove duplication of fatal error reasons

* migrations.maxBatchSizeBytes to docker environment vars

* docs for migrations.maxBatchSizeBytes
# Conflicts:
#	src/core/server/saved_objects/migrationsv2/integration_tests/7_13_0_unknown_types.test.ts
#	src/core/server/saved_objects/migrationsv2/integration_tests/migration_from_v1.test.ts
