+++ src/core/server/saved_objects/migrationsv2/README.md
@@ -316,7 +316,10 @@ completed this step:
  - temp index has a write block
  - temp index is not found
 ### New control state
+1. If `currentBatch` is the last batch in `transformedDocBatches`
   → `REINDEX_SOURCE_TO_TEMP_READ`
+2. If there are more batches left in `transformedDocBatches`
+  → `REINDEX_SOURCE_TO_TEMP_INDEX_BULK`
    
 ## REINDEX_SOURCE_TO_TEMP_CLOSE_PIT
 ### Next action
