+++ src/core/server/saved_objects/migrationsv2/integration_tests/cleanup.test.ts
@@ -13,7 +13,7 @@ import JSON5 from 'json5';
 import * as kbnTestServer from '../../../../test_helpers/kbn_server';
 import type { Root } from '../../../root';
 
-const logFilePath = Path.join(__dirname, 'cleanup_test.log');
+const logFilePath = Path.join(__dirname, 'cleanup.log');
 
 const asyncUnlink = Util.promisify(Fs.unlink);
 const asyncReadFile = Util.promisify(Fs.readFile);
