+++ src/core/server/saved_objects/migrationsv2/integration_tests/rewriting_id.test.ts
@@ -15,7 +15,7 @@ import type { ElasticsearchClient } from '../../../elasticsearch';
 import { Root } from '../../../root';
 import { deterministicallyRegenerateObjectId } from '../../migrations/core/document_migrator';
 
-const logFilePath = Path.join(__dirname, 'migration_test_kibana.log');
+const logFilePath = Path.join(__dirname, 'rewriting_id.log');
 
 const asyncUnlink = Util.promisify(Fs.unlink);
 async function removeLogFile() {
