+++ src/core/server/saved_objects/migrationsv2/test_helpers/retry.test.ts
@@ -1,0 +1,57 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
+ */
+
+import { retryAsync } from './retry_async';
+
+describe('retry', () => {
+  it('retries throwing functions until they succeed', async () => {
+    let i = 0;
+    await expect(
+      retryAsync(
+        () => {
+          if (i++ < 2) {
+            return Promise.reject(new Error('boom'));
+          } else {
+            return Promise.resolve('done');
+          }
+        },
+        { retryAttempts: 3, retryDelayMs: 1 }
+      )
+    ).resolves.toEqual('done');
+  });
+
+  it('throws if all attempts are exhausted before success', async () => {
+    let attempts = 0;
+    await expect(() =>
+      retryAsync<Error>(
+        () => {
+          attempts++;
+          return Promise.reject(new Error('boom'));
+        },
+        { retryAttempts: 3, retryDelayMs: 1 }
+      )
+    ).rejects.toMatchInlineSnapshot(`[Error: boom]`);
+    expect(attempts).toEqual(3);
+  });
+
+  it('waits retryDelayMs between each attempt ', async () => {
+    const now = Date.now();
+    let i = 0;
+    await retryAsync(
+      () => {
+        if (i++ < 2) {
+          return Promise.reject(new Error('boom'));
+        } else {
+          return Promise.resolve('done');
+        }
+      },
+      { retryAttempts: 3, retryDelayMs: 100 }
+    );
+    expect(Date.now() - now).toBeGreaterThanOrEqual(200);
+  });
+});
